+++ goback.hpp
@@ -12,7 +12,7 @@ public:
 	uint64_t white_bit_mask, black_bit_mask, castlingMap, hash;
 	uint8_t moveNumber, ruleNumber, passant_x, passant_y;
 	bool whiteMove, passant_enable;
-	double evalute;
+	int64_t evalute;
 	bool attacked;
 	double margin;
 };
